requirements.txt
streamlit
pymupdf
python-dotenv
apify-client
openrouter
requests
python-magic